Update project version and modify client state handling
In pyproject.toml, updated the version from 0.1.0 to 0.3.0 to keep track of significant updates. In client.py, refined the __setitem__ method in the State class for a more direct handling of state updates, thereby avoiding potential issues with overwriting states. Also, change the endpoint URLs in job creation from '/create/' to '/run/' to better reflect their actual operations. This change was necessary to prevent confusion and improve the readability of the URLs.