Bump project version and remove State class in client
The changes are two-fold: The version in `pyproject.toml` has been incremented to "0.5.0" to mark the completion of significant updates; Concurrently, in `windmill_client/client.py`, the 'State' class, previously used for managing client state, has been removed. The goal of these modifications is to simplify the client's state handling. Instead of maintaining the state with a dedicated class, we now directly return the object received from 'get_resource' method. This makes the state management more transparent and less complex.