Add sync script running instructions to README.md; Increment version
The README.md file has been updated with instructions on how to run a script synchronously using the Python client. This addition should guide users who want to execute scripts without waiting for completion. Additionally, the project version in the pyproject.toml file has been incremented from 0.5.0 to 0.5.1 to denote the addition of this new feature in the client.