Bump to version, add raise_for_status option in HTTP methods
The client's HTTP methods (get, post) have been updated to raise an exception if the return status code from the server is erroneous. A "raise_for_status" parameter is added to toggle this behavior - Resource and variable methods no longer raise an exception if the resource or variable doesn't exist. Simultaneously, the project version is incremented to denote these changes.