Update and expose version, set dynamic version in pyproject.toml
Changed the version of the package to 0.7.1 in windmill_client/__init__.py and marked the version as dynamic in pyproject.toml file. The purpose of exposing the version is to make the current version of the package easily accessible at runtime. Also, by setting the version as dynamic in the pyproject.toml file, it allows the use of version control systems to manage and automatically update the project version.